+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * [#580](https://github.com/mozilla/glean.js/pull/580): BUGFIX: Pending pings at startup up are uploaded from oldest to newest.
 * [#607](https://github.com/mozilla/glean.js/pull/607): Record an error when incoherent timestamps are calculated for events after a restart.
 * [#630](https://github.com/mozilla/glean.js/pull/630): Accept booleans and numbers as event extras.
+* [#658](https://github.com/mozilla/glean.js/pull/658): Implement rate limiting for ping upload.
+  * Only up to 15 ping submissions every 60 seconds are now allowed.
+* [#658](https://github.com/mozilla/glean.js/pull/658): BUGFIX: Unblock ping uploading jobs after the maximum of upload failures are hit for a given uploading window.
 
 # v0.18.1 (2021-07-22)